ArvoreDeDecisão separada em módulos/ Nova lógica para entendimento do contexto adicionada
